+++ tests/CartTest.php
@@ -1,4 +1,4 @@
-ph<?php
+<?php
 
 use Cart\Cart;
 use Cart\CartItem;